Fixes and new file structure
