+++ app/dashboard/components/feedToggle.tsx
@@ -1,29 +1,47 @@
-export default function FeedToggle({ feedType, setFeedType }) {
-  return (
-    <div className="space-y-4 w-full max-w-md mx-auto">
-      {/* Feed Toggle Buttons */}
-      <div className="flex justify-center bg-gray-100 p-1 rounded-lg">
-        <button
-          className={`px-4 py-1.5 rounded-md transition-all duration-200 font-medium ${
-            feedType === "engagements"
-              ? "bg-blue-500 text-white shadow-md"
-              : "bg-transparent text-gray-600 hover:bg-gray-200"
-          }`}
-          onClick={() => setFeedType("engagements")}
-        >
-          My Engagements
-        </button>
-        <button
-          className={`px-4 py-1.5 rounded-md transition-all duration-200 font-medium ${
-            feedType === "feed"
-              ? "bg-blue-500 text-white shadow-md"
-              : "bg-transparent text-gray-600 hover:bg-gray-200"
-          }`}
-          onClick={() => setFeedType("feed")}
-        >
-          My Feed
-        </button>
-      </div>
-    </div>
-  );
+"use client";
+import Link from "next/link";
+import { usePathname, useSearchParams } from "next/navigation";
+
+interface FeedToggleProps {
+    currentFeedType: string;
+}
+
+export default function FeedToggle({ currentFeedType }: FeedToggleProps) {
+    const pathname = usePathname();
+    const searchParams = useSearchParams();
+
+    // Create a new URLSearchParams instance to preserve other parameters
+    const createQueryString = (feedType: string) => {
+        const params = new URLSearchParams(searchParams.toString());
+        params.set("feedType", feedType);
+        return params.toString();
+    };
+
+    return (
+        <div className="space-y-4 w-full max-w-md mx-auto">
+            {/* Feed Toggle Buttons */}
+            <div className="flex justify-center bg-gray-100 p-1 rounded-lg">
+                <Link
+                    href={`${pathname}?${createQueryString("engagements")}`}
+                    className={`px-4 py-1.5 rounded-md transition-all duration-200 font-medium ${
+                        currentFeedType === "engagements"
+                            ? "bg-blue-500 text-white shadow-md"
+                            : "bg-transparent text-gray-600 hover:bg-gray-200"
+                    }`}
+                >
+                    My Engagements
+                </Link>
+                <Link
+                    href={`${pathname}?${createQueryString("feed")}`}
+                    className={`px-4 py-1.5 rounded-md transition-all duration-200 font-medium ${
+                        currentFeedType === "feed"
+                            ? "bg-blue-500 text-white shadow-md"
+                            : "bg-transparent text-gray-600 hover:bg-gray-200"
+                    }`}
+                >
+                    My Feed
+                </Link>
+            </div>
+        </div>
+    );
 }
